refactor: 중복 로직 컴포넌트 분리
