+++ styles/post/post.module.scss
@@ -47,9 +47,9 @@
   align-items: center;
   img {
     border-radius: 50%;
-    margin-right: 12px;
   }
   .nickname {
+    margin-left: 12px;
     font-weight: bold;
   }
 }
